Add modules and refactor code to es6
